02476_FinalProject_Group20/requirements.txt
invoke==2.2.0
Pillow==11.1.0
torch>=2.0
torchvision>=0.17.2
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
